Rewrite layout to use a nested layout
Changed the layout so now when you go into a campaign, the campaignId is at the top of the layout and all subroutes go under. This allows the main layout to only be loaded once and the campaignId to always be available. It also means we never have to copy the layout in order to work out the campaign from its own ID or from the domain being routed on.